demo/requirements.txt
# Your app requirements.
-r ../requirements_test.txt

# Your app in editable mode.
-e ../

django-debug-toolbar==1.9.1
